requirements.txt
Flask==3.0.3
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
flask-mail==0.9.1
itsdangerous==2.2.0
openai==0.27.0
python-dotenv==1.0.0
requests==2.28.2
